+++ cisco_sdwan/base/catalog.py
@@ -52,6 +52,7 @@ _op_catalog = dict()  # {<OpType>: [(<tag>, <selector>, <info>, <op_cls>, <min_v
 # Order in which config items need to be deleted (i.e. reverse order in which they need to be pushed), considering
 # their high-level dependencies.
 _tag_dependency_list = [
+    'rule',
     'config_group',
     'topology_group',
     'policy_group',